cart.py debug decimal

+++ cart/cart.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from django.conf import settings
 from shop.models import Product
 
@@ -30,7 +29,7 @@ class Cart(object):
             cart[str(product.id)]['product'] = product
 
         for item in cart.values():
-            item['price'] = Decimal(item['price'])
+            item['price'] = (item['price'])
             item['total_price'] = item['price'] * item['quantity']
             yield item
 
@@ -73,4 +72,4 @@ class Cart(object):
         self.save()
 
     def get_total_price(self):
-        return sum(Decimal(item['price']) * item['quantity'] for item in self.cart.values())
+        return sum(float(item['price']) * item['quantity'] for item in self.cart.values())
